+++ lib/funny.js
@@ -1,4 +1,5 @@
 import Imgflipper from 'imgflipper';
+import gizoogle from 'gizoogle';
 import nconf from 'nconf';
 
 import sentry from './config/sentry';
@@ -52,10 +53,22 @@ function meme(bot, msg, suffix) {
   });
 }
 
+function snoopify(bot, msg, suffix) {
+  if (!suffix) {
+    bot.sendMessage(msg.channel, 'Usage: **`!snoopify`** `sentence`');
+    return;
+  }
+  gizoogle.string(suffix, (err, translation) => {
+    if (err) sentry.captureError(err);
+    bot.sendMessage(msg.channel, translation);
+  });
+}
+
 export default {
   drama,
   emoji,
   emojis: emoji,
   meme,
-  memes: meme
+  memes: meme,
+  snoopify
 };
